Modifications to get forceworkbench on packagist
I'd like to do ```composer install forceworkbench/forceworkbench``` , but that requires that there is a name field in the composer.json and that this repo is submitted to https://packagist.org/packages/submit .   Would you please approve this change?  I think this will help adoption as I believe this is how most PHP developers add packages.