Adds build commands to run redis on the local build
Adds 3 run commands to the Dockerfile that download and run a local redis Server. Adds commands to the docker-entrypoint.sh that sends the running redis server to the background of the docker shell. Adds values to the overrrides.php file for the redisURL variable to point to the local redis server URL.