Handle the case of unitialized array
See
http://php.net/manual/en/language.types.array.php#language.types.array.syntax.modifying

> As of PHP 7.1.0, applying the empty index operator on a string throws
> a fatal error. Formerly, the string was silently converted to an
> array.

This first manifested in the environment variables in app.json. In
particular, forceworkbench__streamingAdviceTimeout__default since there
was no default setting for streamingAdviceTimeout in defaults.php.

The old way of running this locally didn't read the app.json variables
so we won't see this issue until we are running on Heroku. I've changed
it so that we now run the variables.

There are two variables, forceworkbench__csrfSecret__default and
forceworkbench__rc4Secret__default which we won't read since it relies
on `generator: secret` in app.json.